Split KolibriWindow into a separate file
